FlaskHelper/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyPDF2==3.0.1
youtube-transcript-api==0.6.2
langchain==0.0.201
pinecone-client==5.0.0
google-generativeai==0.1.0
requests==2.31.0
